feat: refactor video recording
