+++ README.md
@@ -7,6 +7,9 @@
 - [x] Fix clerk sign up
 - [ ] Refactor server files in to a server folder (e.g. server/db/schema)
 - [ ] Isolate the video render between recording video stream and playback. After recording "No video recording" message is briefly shown
+- [ ] Add toast component for error handling inside handleStartRecording function
+- [ ] Get a large fixed store of table topics (Will eventually be evolved to use AI generated topics)
+- [ ] remove naming feature for the video download
 
 ## 1. Project Setup
