+++ app/topic/create/page.tsx
@@ -1,9 +1,5 @@
-import TableTopicsRecorder from "@/components/table-topics-recorder";
+import TableTopicsRecorder from "./_components/table-topics-recorder";
 
 export default function TopicCreatePage() {
-  return (
-    <>
-      <TableTopicsRecorder />
-    </>
-  );
+  return <TableTopicsRecorder />;
 }
